backend/password-reset/requirements.txt
psycopg2-binary==2.9.9
PyJWT==2.8.0
bcrypt==4.1.2